Refactor from type User to interface DBEntry Add type UserStatus to types.d.ts Move data.ts to /utils
